+++ src/lwip_driver.h
@@ -67,9 +67,9 @@ extern "C" {
 // Returns if the hardware hasn't yet been probed.
 bool driver_is_unknown();
 
-// Returns whether the link state is unknown. This will return true if the
-// hardware is not capable of detecting a link and false otherwise.
-bool driver_is_link_state_unknown();
+// Returns whether the link state is detectable. This will return true if the
+// hardware is capable of detecting a link and false otherwise.
+bool driver_is_link_state_detectable();
 
 // Gets the built-in Ethernet MAC address.
 //